Add files via upload
Added support for episode pages. The script now scans for episode pages, and creates a link for that specific episode.

Works by scanning for
aria-label="View all episodes"
which only exist on episode pages.
It then finds the main series link, slaps on the episode number and adds the button. Very elegant.

Same changes to IMDb AutoRePo, in addition to version hopping all the way up to 6.0 for parity.